+++ lib/vagrant-notify-forwarder/pid_handler.rb
@@ -1,0 +1,14 @@
+require 'singleton'
+
+module VagrantPlugins
+  module VagrantNotifyForwarder
+    class PidHandler
+      include Singleton
+      attr_accessor :pids
+
+      def initialize
+        @pids = []
+      end
+    end
+  end
+end
